修复温度计算溢出 手柄触发选择开关 更好的温度拟合常数 保存图标

+++ TipControl.h
@@ -2,8 +2,6 @@
 #define TipControl_H
 #include "Type.h"
 
-
-
 extern uint8_t MyMOS;
 extern uint8_t PWM;
 extern uint16_t LastADC;
@@ -18,7 +16,7 @@ extern double consKp, consKi, consKd;
 void TipControlInit(void);
 void PWMOutput(uint8_t pin, uint8_t pwm);
 int GetADC0(void);
-double CalculateTemp(uint16_t ADC);
+double CalculateTemp(double ADC);
 void SetPOWER(uint8_t POWER);
 
 void TemperatureControlLoop(void);
